# docs/openapi/knowledge.openapi.yaml
openapi: 3.1.0
info:
  title: knowledge
  version: 1.0.0
  description: knowledge 控制面、知识生产、发布、检索与 Agent Run API。
paths:
...
  /warehouse/status:
    get:
      tags:
      - warehouse_sources
      summary: Warehouse Status
      operationId: warehouse_status_warehouse_status_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WarehouseStatusResponse'
      security:
      - HTTPBearer: []
  /warehouse/bootstrap/challenge:
    post:
      tags:
      - warehouse_sources
      summary: Warehouse Bootstrap Challenge
      operationId: warehouse_bootstrap_challenge_warehouse_bootstrap_challenge_post
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WarehouseBootstrapChallengeResponse'
      security:
      - HTTPBearer: []
  /warehouse/bootstrap/initialize:
    post:
      tags:
      - warehouse_sources
      summary: Warehouse Bootstrap Initialize
      operationId: warehouse_bootstrap_initialize_warehouse_bootstrap_initialize_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/WarehouseBootstrapInitializeRequest'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WarehouseBootstrapInitializeResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      security:
      - HTTPBearer: []
  /warehouse/credentials/read:
    get:
      tags:
      - warehouse_sources
      summary: List Read Credentials
      operationId: list_read_credentials_warehouse_credentials_read_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/WarehouseCredentialSummary'
                type: array
                title: Response List Read Credentials Warehouse Credentials Read Get
      security:
      - HTTPBearer: []
    post:
      tags:
      - warehouse_sources
      summary: Create Read Credential
      operationId: create_read_credential_warehouse_credentials_read_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/WarehouseCredentialCreateRequest'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WarehouseCredentialSummary'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      security:
      - HTTPBearer: []
  /warehouse/credentials/read/{credential_id}/secret:
    get:
      tags:
      - warehouse_sources
      summary: Reveal Read Credential Secret
      operationId: reveal_read_credential_secret_warehouse_credentials_read__credential_id__secret_get
      security:
      - HTTPBearer: []
      parameters:
      - name: credential_id
        in: path
        required: true
        schema:
          type: integer
          title: Credential Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WarehouseCredentialRevealResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /warehouse/credentials/read/{credential_id}:
    delete:
      tags:
      - warehouse_sources
      summary: Delete Read Credential
      operationId: delete_read_credential_warehouse_credentials_read__credential_id__delete
      security:
      - HTTPBearer: []
      parameters:
      - name: credential_id
        in: path
        required: true
        schema:
          type: integer
          title: Credential Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                title: Response Delete Read Credential Warehouse Credentials Read  Credential Id  Delete
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /warehouse/credentials/write:
    get:
      tags:
      - warehouse_sources
      summary: Get Write Credential
      operationId: get_write_credential_warehouse_credentials_write_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WarehouseWriteCredentialResponse'
      security:
      - HTTPBearer: []
    post:
      tags:
      - warehouse_sources
      summary: Upsert Write Credential
      operationId: upsert_write_credential_warehouse_credentials_write_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/WarehouseCredentialCreateRequest'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WarehouseCredentialSummary'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      security:
      - HTTPBearer: []
    delete:
      tags:
      - warehouse_sources
      summary: Delete Write Credential
      operationId: delete_write_credential_warehouse_credentials_write_delete
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                title: Response Delete Write Credential Warehouse Credentials Write Delete
      security:
      - HTTPBearer: []
  /warehouse/credentials/write/secret:
    get:
      tags:
      - warehouse_sources
      summary: Reveal Write Credential Secret
      operationId: reveal_write_credential_secret_warehouse_credentials_write_secret_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WarehouseCredentialRevealResponse'
      security:
      - HTTPBearer: []
  /warehouse/browse:
    get:
      tags:
      - warehouse_sources
      summary: Browse Warehouse
      operationId: browse_warehouse_warehouse_browse_get
      security:
      - HTTPBearer: []
      parameters:
      - name: path
        in: query
        required: false
        schema:
          type: string
          default: /apps/knowledge.yeying.pub
          title: Path
      - name: credential_id
        in: query
        required: false
        schema:
          anyOf:
          - type: integer
          - type: 'null'
          title: Credential Id
      - name: use_write_credential
        in: query
        required: false
        schema:
          type: boolean
          default: false
          title: Use Write Credential
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WarehouseBrowseResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /warehouse/upload:
    post:
      tags:
      - warehouse_sources
      summary: Upload To App Dir
      operationId: upload_to_app_dir_warehouse_upload_post
      requestBody:
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/Body_upload_to_app_dir_warehouse_upload_post'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UploadResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      security:
      - HTTPBearer: []
  /warehouse/uploads:
    get:
      tags:
      - warehouse_sources
      summary: List Upload Records
      operationId: list_upload_records_warehouse_uploads_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/UploadRecordRead'
                type: array
                title: Response List Upload Records Warehouse Uploads Get
      security:
      - HTTPBearer: []
...
components:
  schemas:
...
    WarehouseBootstrapChallengeResponse:
      properties:
        wallet_address:
          type: string
          title: Wallet Address
        challenge:
          type: string
          title: Challenge
        nonce:
          anyOf:
          - type: string
          - type: 'null'
          title: Nonce
        issued_at:
          anyOf:
          - type: integer
          - type: 'null'
          title: Issued At
        expires_at:
          anyOf:
          - type: integer
          - type: 'null'
          title: Expires At
      type: object
      required:
      - wallet_address
      - challenge
      title: WarehouseBootstrapChallengeResponse
    WarehouseBootstrapInitializeRequest:
      properties:
        mode:
          type: string
          enum:
          - uploads_bundle
          - app_root_write
          title: Mode
          default: uploads_bundle
        signature:
          type: string
          title: Signature
      type: object
      required:
      - signature
      title: WarehouseBootstrapInitializeRequest
...
    WarehouseStatusResponse:
      properties:
        wallet_address:
          type: string
          title: Wallet Address
        credentials_ready:
          type: boolean
          title: Credentials Ready
          default: false
        read_credentials_count:
          type: integer
          title: Read Credentials Count
          default: 0
        write_credential_id:
          anyOf:
          - type: integer
          - type: 'null'
          title: Write Credential Id
        write_credential_status:
          anyOf:
          - type: string
          - type: 'null'
          title: Write Credential Status
        write_root_path:
          anyOf:
          - type: string
          - type: 'null'
          title: Write Root Path
        current_app_id:
          anyOf:
          - type: string
          - type: 'null'
          title: Current App Id
        current_app_root:
          anyOf:
          - type: string
          - type: 'null'
          title: Current App Root
        current_app_upload_dir:
          anyOf:
          - type: string
          - type: 'null'
          title: Current App Upload Dir
        warehouse_base_url:
          anyOf:
          - type: string
          - type: 'null'
          title: Warehouse Base Url
      type: object
      required:
      - wallet_address
      title: WarehouseStatusResponse
...
  securitySchemes:
    HTTPBearer:
      type: http
      scheme: bearer
servers:
- url: http://127.0.0.1:8000
  description: 本地开发环境